♻️ Refactor : 랜딩 화면 로직만 MainPage.tsx로 분리, Logo 컴포넌트로 분리
